Removed dependency hell
Previously, the DataHolder base class was in param.py, whilst the main
DataHolders were (sensibly) in data_holders.py. This was due to some
impressive dependency hell. To fix, I moved the Parentable base class
out into tree_structure.py. I also modified slightly the PAramList
object, so that it's less of a 'special case' and doesn't require
improting into tree_stuctures. This improves the code by reducing
circularity.
