+++ GPflow/__init__.py
@@ -1,3 +1,3 @@
 # flake8: noqa
-from . import likelihoods, kernels, param, model, gpmc, sgpmc, priors, gpr, svgp, vgp, sgpr
+from . import likelihoods, kernels, param, model, gpmc, sgpmc, priors, gpr, svgp, vgp, sgpr, mean_functions, data_holders
 from ._version import __version__
