+++ GPflow/tree_structure.py
@@ -1,0 +1,53 @@
+class Parentable(object):
+    """
+    A very simple class for objects in a tree, where each node contains a
+    reference to '_parent'.
+
+    This class can figure out its own name (by seeing what it's called by the
+    _parent's __dict__) and also recurse up to the highest_parent.
+    """
+
+    def __init__(self):
+        self._parent = None
+
+    @property
+    def highest_parent(self):
+        if self._parent is None:
+            return self
+        else:
+            return self._parent.highest_parent
+
+    @property
+    def name(self):
+        """to get the name of this object, have a look at
+        what our _parent has called us"""
+        if self._parent is None:
+            return 'unnamed'
+        matches = [key for key, value in self._parent.__dict__.items()
+                   if value is self]
+        if len(matches) == 0:
+            raise ValueError("mis-specified parent. This Param's\
+                             _parent does not contain a reference to it.")
+        if len(matches) > 1:
+            raise ValueError("This Param appears to be doubly\
+                             referenced by a parent")
+        return matches[0]
+
+    @property
+    def long_name(self):
+        """
+        This is a unique identifier for a param object within a structure, made
+        by concatenating the names through the tree.
+        """
+        if self._parent is None:
+            return self.name
+        return self._parent.long_name + '.' + self.name
+
+    def __getstate__(self):
+        d = self.__dict__.copy()
+        d.pop('_parent')
+        return d
+
+    def __setstate__(self, d):
+        self.__dict__.update(d)
+        self._parent = None
